+++ src/main/java/randommcsomethin/fallingleaves/LeafUtils.java
@@ -6,50 +6,49 @@ import net.minecraft.client.texture.Sprite;
 import net.minecraft.particle.DefaultParticleType;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
-import randommcsomethin.fallingleaves.config.OverrideEntry;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import randommcsomethin.fallingleaves.config.LeafSettingsEntry;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
-import java.util.Objects;
 
 import static randommcsomethin.fallingleaves.init.Config.CONFIG;
 
 public class LeafUtils {
 
-    static boolean DEBUG = CONFIG.displayDebugData;
-
-    public static void debugLog(String debugdata) {
-        if (LeafUtils.DEBUG) {
-            System.out.println("[fallingleaves] " + debugdata);
-        }
-    }
+    public static final Logger LOGGER = LogManager.getLogger("fallingleaves");
 
     public static double getLeafSpawnRate(BlockState state) {
         String blockId = getBlockId(state);
+        LeafSettingsEntry leafSettingsEntry = getLeafSettingsEntry(blockId);
 
-        OverrideEntry overrideEntry = getOverrideConfigEntry(blockId);
-
-        if (overrideEntry.useCustomSpawnRate) {
-            return overrideEntry.spawnRate;
+        if (leafSettingsEntry == null) {
+            LOGGER.warn("There is no config entry for {}!", blockId); // TODO: see LeafSettings.validatePostLoad()
+            return 1;
         } else {
-            return isConifer(blockId) ? CONFIG.coniferLeafSpawnRate : CONFIG.leafSpawnRate;
+            if (leafSettingsEntry.useCustomSpawnRate) {
+                return leafSettingsEntry.spawnRate;
+            } else {
+                return leafSettingsEntry.isConiferBlock ? CONFIG.coniferLeafSpawnRate : CONFIG.leafSpawnRate;
+            }
         }
     }
 
-    private static OverrideEntry getOverrideConfigEntry(String blockId) {
-        return CONFIG.override.overrideEntries
-            .stream()
-            .filter(blocks -> Objects.equals(blocks.identifier, blockId))
-            .findFirst()
-            .orElse(null);
+    public static boolean isConifer(BlockState state) {
+        String blockId = getBlockId(state);
+        LeafSettingsEntry leafSettingsEntry = getLeafSettingsEntry(blockId);
+
+        if (leafSettingsEntry == null) {
+            LOGGER.warn("There is no config entry for {}!", blockId); // TODO: see LeafSettings.validatePostLoad()
+            return false;
+        } else {
+            return leafSettingsEntry.isConiferBlock;
+        }
     }
 
-    public static boolean isConifer(BlockState blockState) {
-        return isConifer(getBlockId(blockState));
-    }
-
-    public static boolean isConifer(String block) {
-        return getOverrideConfigEntry(block).isConiferBlock;
+    private static LeafSettingsEntry getLeafSettingsEntry(String blockId) {
+        return CONFIG.leafSettings.entries.get(new LeafSettingsEntry(new Identifier(blockId)));
     }
 
     public static String getBlockId(BlockState state) {
@@ -57,7 +56,7 @@ public class LeafUtils {
     }
 
     public static DefaultParticleType registerParticle(String id) {
-        LeafUtils.debugLog("Registering particle: " + id);
+        LOGGER.debug("Registering particle: {}", id);
         return Registry.register(Registry.PARTICLE_TYPE, makeIdent(id), FabricParticleTypes.simple());
     }
 
@@ -74,10 +73,10 @@ public class LeafUtils {
         // TODO: This entire block feels like it could be simplified.
         for (int y = 0; y < image.getHeight(); y++) {
             for (int x = 0; x < image.getWidth(); x++) {
-                int rgb = image.getRGB(x, y);
+                Color c = new Color(image.getRGB(x, y), true);
 
-                if ((rgb >> 24 & 255) == 255) {
-                    Color c = new Color(rgb);
+                // Only take completely opaque pixels into account
+                if (c.getAlpha() == 255) {
                     r += c.getRed();
                     g += c.getGreen();
                     b += c.getBlue();
@@ -87,9 +86,7 @@ public class LeafUtils {
             }
         }
 
-        LeafUtils.debugLog("Generated new average color. RGB: " + r / n + "/" + g / n + "/" + b / n);
-
-        return new Color((float) r / n / 255, (float) g / n / 255, (float) b / n / 255);
+        return new Color((int)(r / n), (int)(g / n), (int)(b / n));
     }
 
     public static String spriteToTexture(Sprite sprite) {