+++ src/main/java/randommcsomethin/fallingleaves/TextureCache.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
 
 public class TextureCache {
     public static class Data {
-        public int color;
-        public String resourcePack;
+        public final int color;
+        public final String resourcePack;
 
         public Data(int color, String resourcePack) {
             this.color = color;